creating ts files
