+++ src/app/constants/filters.ts
@@ -1,0 +1,1 @@
+export const filters = ["All", "Frontend", "Backend", "HR", "AI"];
