+++ src/app/constants/testimonials.ts
@@ -1,0 +1,20 @@
+export const testimonials = [
+  {
+    name: "Ahmet K.",
+    role: "Frontend Developer",
+    text: "SpeakHire sayesinde dream job'ımı aldım! Mülakat pratiği harika.",
+    rating: 5,
+  },
+  {
+    name: "Zeynep M.",
+    role: "Backend Engineer",
+    text: "AI destekli feedback sistemi gerçekten işe yarıyor. Kendimi çok geliştirdim.",
+    rating: 5,
+  },
+  {
+    name: "Can S.",
+    role: "Full Stack Developer",
+    text: "Gerçek mülakat ortamını çok iyi simüle ediyor. Tavsiye ederim!",
+    rating: 5,
+  },
+];
