+++ src/app/page.tsx
@@ -9,17 +9,9 @@ import {
   Grid,
   useTheme,
   Card,
-  Stack,
+  Stack
 } from "@mui/material";
-import {
-  Search,
-  ChevronRight,
-  PlayArrow,
-  EmojiEvents,
-  Star,
-  Person,
-} from "@mui/icons-material";
-import { FaCode, FaDatabase, FaUsers, FaBrain, FaDocker } from "react-icons/fa";
+import { Search, ChevronRight } from "@mui/icons-material";
 import { SearchBar } from "@/components/SearchBar";
 import { AlertBox } from "@/components/AlertBox";
 import { FilterButtons } from "@/components/FilterButtons";
@@ -27,6 +19,10 @@ import { CardStructure } from "@/components/CardStructure";
 import { Footer } from "@/components/Footer/Footer";
 import { motion, AnimatePresence } from "framer-motion";
 import { Testimonials } from "@/components/Testimonials";
+import { interviewCards } from "./constants/interviewCards";
+import { testimonials } from "./constants/testimonials";
+import { stats } from "./constants/stats";
+import { filters } from "./constants/filters";
 
 export default function Home() {
   const [activeFilter, setActiveFilter] = useState("All");
@@ -34,121 +30,6 @@ export default function Home() {
   const theme = useTheme();
   const [searchTerm, setSearchTerm] = useState("");
   const [currentTestimonial, setCurrentTestimonial] = useState(0);
-
-  const filters = ["All", "Frontend", "Backend", "HR", "AI"];
-
-  const interviewCards = [
-    {
-      id: 1,
-      title: "Frontend Developer",
-      description: "React, JavaScript, CSS, HTML konularında sorular içerir.",
-      difficulty: "Beginner",
-      category: "Frontend",
-      icon: FaCode,
-      duration: "45 min",
-      questions: 25,
-      rating: 4.8,
-    },
-    {
-      id: 2,
-      title: "Backend Developer",
-      description: "Node.js, Express.js, SQL konularında sorular içerir.",
-      difficulty: "Intermediate",
-      category: "Backend",
-      icon: FaDatabase,
-      duration: "60 min",
-      questions: 30,
-      rating: 4.7,
-    },
-    {
-      id: 3,
-      title: "HR Mülakatı",
-      description: "Davranışsal sorular, ekip çalışması ve liderlik.",
-      difficulty: "Beginner",
-      category: "HR",
-      icon: FaUsers,
-      duration: "30 min",
-      questions: 20,
-      rating: 4.9,
-    },
-    {
-      id: 4,
-      title: "Full Stack Developer",
-      description: "Frontend ve Backend teknolojilerinin tamamı.",
-      difficulty: "Advanced",
-      category: "FullStack",
-      icon: FaCode,
-      duration: "90 min",
-      questions: 45,
-      rating: 4.6,
-    },
-    {
-      id: 5,
-      title: "AI/ML Engineer",
-      description: "Machine Learning, Deep Learning, Python.",
-      difficulty: "Advanced",
-      category: "AI",
-      icon: FaBrain,
-      duration: "75 min",
-      questions: 35,
-      rating: 4.8,
-    },
-    {
-      id: 6,
-      title: "DevOps Engineer",
-      description: "Docker, Kubernetes, CI/CD, AWS konuları.",
-      difficulty: "Intermediate",
-      category: "DevOps",
-      icon: FaDocker,
-      duration: "50 min",
-      questions: 28,
-      rating: 4.7,
-    },
-  ];
-
-  const testimonials = [
-    {
-      name: "Ahmet K.",
-      role: "Frontend Developer",
-      text: "SpeakHire sayesinde dream job'ımı aldım! Mülakat pratiği harika.",
-      rating: 5,
-    },
-    {
-      name: "Zeynep M.",
-      role: "Backend Engineer",
-      text: "AI destekli feedback sistemi gerçekten işe yarıyor. Kendimi çok geliştirdim.",
-      rating: 5,
-    },
-    {
-      name: "Can S.",
-      role: "Full Stack Developer",
-      text: "Gerçek mülakat ortamını çok iyi simüle ediyor. Tavsiye ederim!",
-      rating: 5,
-    },
-  ];
-
-  const stats = [
-    {
-      label: "Active Users",
-      value: "50K+",
-      icon: <Person sx={{ fontSize: 24 }} />,
-    },
-    {
-      label: "Interview Sessions",
-      value: "200K+",
-      icon: <PlayArrow sx={{ fontSize: 24 }} />,
-    },
-    {
-      label: "Success Rate",
-      value: "85%",
-      icon: <EmojiEvents sx={{ fontSize: 24 }} />,
-    },
-    {
-      label: "Avg. Rating",
-      value: "4.8★",
-      icon: <Star sx={{ fontSize: 24 }} />,
-    },
-  ];
 
   const filteredCards = interviewCards.filter((card) => {
     const matchesSearch =
@@ -178,7 +59,6 @@ export default function Home() {
     });
   }, [theme]);
 
-  // Auto-rotate testimonials
   useEffect(() => {
     const timer = setInterval(() => {
       setCurrentTestimonial((prev) => (prev + 1) % testimonials.length);
@@ -241,7 +121,8 @@ export default function Home() {
                     maxWidth: "600px",
                   }}
                 >
-                  Practice with realistic scenarios, get instant feedback, and land your dream job with confidence.
+                  Practice with realistic scenarios, get instant feedback, and
+                  land your dream job with confidence.
                 </Typography>
                 <Stack direction={{ xs: "column", sm: "row" }} spacing={2}>
                   <Button
@@ -280,28 +161,33 @@ export default function Home() {
                 transition={{ duration: 0.8, delay: 0.2 }}
               >
                 <Grid container spacing={2}>
-                  {stats.map((stat, index) => (
-                    <Grid container key={index}>
-                      <Card
-                        sx={{
-                          background: "rgba(255,255,255,0.15)",
-                          backdropFilter: "blur(10px)",
-                          border: "1px solid rgba(255,255,255,0.2)",
-                          color: "white",
-                          textAlign: "center",
-                          p: 2,
-                        }}
-                      >
-                        <Box sx={{ color: "#FFD700", mb: 1 }}>{stat.icon}</Box>
-                        <Typography variant="h4" fontWeight="bold">
-                          {stat.value}
-                        </Typography>
-                        <Typography variant="body2" sx={{ opacity: 0.8 }}>
-                          {stat.label}
-                        </Typography>
-                      </Card>
-                    </Grid>
-                  ))}
+                  {stats.map((stat, index) => {
+                    const Icon = stat.icon;
+                    return (
+                      <Grid container key={index}>
+                        <Card
+                          sx={{
+                            background: "rgba(255,255,255,0.15)",
+                            backdropFilter: "blur(10px)",
+                            border: "1px solid rgba(255,255,255,0.2)",
+                            color: "white",
+                            textAlign: "center",
+                            p: 2,
+                          }}
+                        >
+                          <Box sx={{ color: "#FFD700", mb: 1 }}>
+                            <Icon sx={{ fontSize: 24 }} />
+                          </Box>
+                          <Typography variant="h4" fontWeight="bold">
+                            {stat.value}
+                          </Typography>
+                          <Typography variant="body2" sx={{ opacity: 0.8 }}>
+                            {stat.label}
+                          </Typography>
+                        </Card>
+                      </Grid>
+                    );
+                  })}
                 </Grid>
               </motion.div>
             </Grid>
@@ -354,11 +240,17 @@ export default function Home() {
 
         {/* Enhanced Interview Cards */}
         <Box ref={cardsRef}>
-          <Grid container spacing={3} columns={3} alignItems={"center"} justifyContent={"center"} >
+          <Grid
+            container
+            spacing={3}
+            columns={3}
+            alignItems={"center"}
+            justifyContent={"center"}
+          >
             {filteredCards.slice(0, 6).map((card, index) => {
               const IconComponent = card.icon;
               return (
-                <Grid container key={card.id} >
+                <Grid container key={card.id}>
                   <motion.div
                     initial={{ opacity: 0, y: 30 }}
                     animate={{ opacity: 1, y: 0 }}
@@ -432,5 +324,3 @@ export default function Home() {
     </Box>
   );
 }
-
-// useMemo ve useCallBack nedir ? Teorik kısmını öğren. Syntax kısmına bak.
